refactor: domain name; add support for notification type; reorganize structure
